Fix FireBaseDBContext

+++ AlphaProject/Libraries/FireBaseDB/DB/Impl/FireBaseDBContext.cs
@@ -5,6 +5,7 @@ using FireSharp.Response;
 using ModleLibrary.Model;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using System.Net;
 
 namespace FireBaseDB.DB.Impl
 {
@@ -32,7 +33,7 @@ namespace FireBaseDB.DB.Impl
                 var data = person;
                 SetResponse setResponse = client.Set("Persons/" + data.GuidPerson, data);
 
-                if (setResponse.StatusCode == System.Net.HttpStatusCode.OK)
+                if (setResponse.StatusCode == HttpStatusCode.OK)
                 {
                     return person.GuidPerson;
                 }
@@ -50,41 +51,89 @@ namespace FireBaseDB.DB.Impl
         /// <inheritdoc />
         public List<Person> GetPersonList()
         {
-            FirebaseResponse response = client.Get("Persons");
-            dynamic data = JsonConvert.DeserializeObject<dynamic>(response.Body);
-            var list = new List<Person>();
-            if (data != null)
+            List<Person> list = new List<Person>();
+            try
             {
-                foreach (var item in data)
+                FirebaseResponse response = client.Get("Persons");
+
+                if (response.StatusCode == HttpStatusCode.OK && !string.IsNullOrWhiteSpace(response.Body))
                 {
-                    list.Add(JsonConvert.DeserializeObject<Person>(((JProperty)item).Value.ToString()));
+                    dynamic data = JsonConvert.DeserializeObject<dynamic>(value: response.Body);
+
+                    if (data != null)
+                    {
+                        foreach (var item in data)
+                        {
+                            list.Add(JsonConvert.DeserializeObject<Person>(((JProperty)item).Value.ToString()));
+                        }
+                    }
                 }
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
             }
 
             return list;
         }
 
         /// <inheritdoc />
-        public Person GetPerson(string GuidPerson)
+        public Person? GetPerson(string GuidPerson)
         {
-            FirebaseResponse response = client.Get("Persons/" + GuidPerson);
-            Person data = JsonConvert.DeserializeObject<Person>(response.Body);
+            Person data=new Person();
+
+            try
+            {
+                FirebaseResponse response = client.Get("Persons/" + GuidPerson);
+                if (response.StatusCode == HttpStatusCode.OK && !string.IsNullOrWhiteSpace(response.Body))
+                {
+                    data = JsonConvert.DeserializeObject<Person>(response.Body);
+                }
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
 
             return data;
         }
 
         /// <inheritdoc />
-        public bool UpdatePerson(Person person)
+        public Boolean UpdatePerson(Person person)
         {
-            SetResponse response = client.Set("Persons/" + person.GuidPerson, person);
+            try
+            {
+                if (person == null || string.IsNullOrWhiteSpace(person.GuidPerson))
+                {
+                    throw new ArgumentException("Person or GuidPerson cannot be null or empty.");
+                }
+
+                SetResponse response = client.Set("Persons/" + person.GuidPerson, person);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
 
             return true;
         }
 
         /// <inheritdoc />
-        public bool DeletePerson(string GuidPerson)
+        public Boolean DeletePerson(string GuidPerson)
         {
-            FirebaseResponse response = client.Delete("Persons/" + GuidPerson);
+            try
+            {
+                if (string.IsNullOrWhiteSpace(GuidPerson))
+                {
+                    throw new ArgumentException("GuidPerson cannot be null or empty.");
+                }
+
+                FirebaseResponse response = client.Delete("Persons/" + GuidPerson);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
 
             return true;
         }
